# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Documentação da API para gerenciamento de Usuários, Posts e Tags.",
        "contact": {
            "name": "Equipe de Suporte",
            "email": "suporte@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Listar todos os posts",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "responses": {
                    "200": {
                        "description": "Lista de posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Criar um novo post",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Meu Primeiro Post"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Conteúdo do post."
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Obter um post pelo ID",
                "operationId": "b0f9d92879218f68a376b6edc6467c67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post recuperado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualizar um post",
                "operationId": "929614537d76a081a92170a41ed3bdb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Título Atualizado"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Conteúdo atualizado."
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Excluir um post",
                "operationId": "fad1726bd24cd459fc1525b67985df39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post removido com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post removido com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "title": "Tag",
                "description": "Representa uma tag associada a posts",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laravel"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "description": "Representa um post criado pelo usuário",
                "properties": {
                    "id": {
                        "description": "ID do post",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Título do post",
                        "type": "string",
                        "example": "Meu Primeiro Post"
                    },
                    "content": {
                        "description": "Conteúdo do post",
                        "type": "string",
                        "example": "Este é o conteúdo do post."
                    },
                    "user_id": {
                        "description": "ID do usuário que criou o post",
                        "type": "integer",
                        "example": 1
                    },
                    "tags": {
                        "description": "Lista de tags associadas ao post",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "created_at": {
                        "description": "Data de criação do post",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data da última atualização do post",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T15:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}